fix: Cast numSamples to int in copyFrom

+++ Source/PluginProcessor.cpp
@@ -318,7 +318,7 @@ void PhonkSynthAudioProcessor::processBlock (juce::AudioBuffer<float>& buffer, j
 
         // Create a buffer for reverb processing
         juce::AudioBuffer<float> reverbBuffer (numChannels, numSamples);
-        reverbBuffer.copyFrom (0, 0, buffer, 0, numSamples);
+        reverbBuffer.copyFrom (0, 0, buffer, 0, (int) numSamples);
 
         // Process reverb using JUCE 7 DSP pattern
         auto reverbBlock = juce::dsp::AudioBlock<float> (reverbBuffer);
